Update footer - resolve conversation
