+++ Web/starter_project_vanguard/types/footer-navigation.ts
@@ -1,0 +1,7 @@
+interface LinkItem {
+    title: string
+    links: {
+      name: string
+      path: string
+    }[]
+  }
